+++ website/index.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-$version = '0.8.30.18';
+$version = '0.8.30.19';
 $repository = 'https://github.com/ediril/hook-line-sync';
 $year = (int) date('Y');
 $title = 'Hook Line Sync — a better way to "just FTP it"';